backend/requirements.txt
# OSGB Backend - Python Dependencies

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
asyncpg==0.29.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Validation & Settings
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1

# Utils
httpx==0.27.0
python-dateutil==2.9.0
